+++ .github/whitelist-capital-letters.txt
@@ -4,9 +4,6 @@
 ./apps/browser/src/safari/desktop/Base.lproj
 ./apps/browser/src/services/vaultTimeout
 ./apps/browser/store/windows/Assets
-./apps/desktop/src/models/nativeMessaging
-./apps/desktop/src/models/nativeMessaging/encryptedMessagePayloads
-./apps/desktop/src/models/nativeMessaging/encryptedMessageResponses
 ./libs/common/spec/misc/logInStrategies
 ./libs/common/src/abstractions/fileDownload
 ./libs/common/src/abstractions/userVerification
@@ -186,42 +183,9 @@
 ./apps/desktop/resources/appx/StoreLogo.png
 ./apps/desktop/resources/appx/Wide310x150Logo.png
 ./apps/desktop/resources/appx/Square44x44Logo.png
-./apps/desktop/native-messaging-test-runner/src/ipcService.ts
-./apps/desktop/native-messaging-test-runner/src/nativeMessageService.ts
-./apps/desktop/native-messaging-test-runner/src/logUtils.ts
 ./apps/desktop/README.md
 ./apps/desktop/desktop_native/Cargo.toml
 ./apps/desktop/desktop_native/Cargo.lock
-./apps/desktop/src/app/services/desktopFileDownloadService.ts
-./apps/desktop/src/models/nativeMessaging/unencryptedCommand.ts
-./apps/desktop/src/models/nativeMessaging/encryptedMessage.ts
-./apps/desktop/src/models/nativeMessaging/legacyMessageWrapper.ts
-./apps/desktop/src/models/nativeMessaging/unencryptedMessage.ts
-./apps/desktop/src/models/nativeMessaging/unencryptedMessageResponse.ts
-./apps/desktop/src/models/nativeMessaging/encryptedCommand.ts
-./apps/desktop/src/models/nativeMessaging/encryptedMessageResponse.ts
-./apps/desktop/src/models/nativeMessaging/encryptedMessageResponses/successStatusResponse.ts
-./apps/desktop/src/models/nativeMessaging/encryptedMessageResponses/generateResponse.ts
-./apps/desktop/src/models/nativeMessaging/encryptedMessageResponses/failureStatusResponse.ts
-./apps/desktop/src/models/nativeMessaging/encryptedMessageResponses/cannotDecryptErrorResponse.ts
-./apps/desktop/src/models/nativeMessaging/encryptedMessageResponses/userStatusErrorResponse.ts
-./apps/desktop/src/models/nativeMessaging/encryptedMessageResponses/accountStatusResponse.ts
-./apps/desktop/src/models/nativeMessaging/encryptedMessageResponses/encryptedMessageResponse.ts
-./apps/desktop/src/models/nativeMessaging/encryptedMessageResponses/cipherResponse.ts
-./apps/desktop/src/models/nativeMessaging/decryptedCommandData.ts
-./apps/desktop/src/models/nativeMessaging/messageCommon.ts
-./apps/desktop/src/models/nativeMessaging/legacyMessage.ts
-./apps/desktop/src/models/nativeMessaging/encryptedMessagePayloads/credentialUpdatePayload.ts
-./apps/desktop/src/models/nativeMessaging/encryptedMessagePayloads/passwordGeneratePayload.ts
-./apps/desktop/src/models/nativeMessaging/encryptedMessagePayloads/credentialCreatePayload.ts
-./apps/desktop/src/models/nativeMessaging/encryptedMessagePayloads/credentialRetrievePayload.ts
-./apps/desktop/src/main/desktopCredentialStorageListener.ts
-./apps/desktop/src/main/powerMonitor.main.ts
-./apps/desktop/src/main/nativeMessaging.main.ts
-./apps/desktop/src/services/passwordReprompt.service.ts
-./apps/desktop/src/services/encryptedMessageHandlerService.ts
-./apps/desktop/src/services/nativeMessaging.service.ts
-./apps/desktop/src/services/nativeMessageHandler.service.ts
 ./apps/cli/stores/chocolatey/tools/VERIFICATION.txt
 ./apps/cli/README.md
 ./apps/browser/README.md
