+++ apps/desktop/native-messaging-test-runner/src/commands/bw-credential-update.ts
@@ -5,9 +5,9 @@ import { hideBin } from "yargs/helpers";
 
 import { NativeMessagingVersion } from "@bitwarden/common/enums/nativeMessagingVersion";
 
-import { CredentialUpdatePayload } from "../../../src/models/nativeMessaging/encryptedMessagePayloads/credentialUpdatePayload";
-import { LogUtils } from "../logUtils";
-import NativeMessageService from "../nativeMessageService";
+import { CredentialUpdatePayload } from "../../../src/models/native-messaging/encrypted-message-payloads/credential-update-payload";
+import { LogUtils } from "../log-utils";
+import NativeMessageService from "../native-message.service";
 import * as config from "../variables";
 
 // Command line arguments
