+++ apps/desktop/native-messaging-test-runner/src/commands/bw-generate-password.ts
@@ -5,8 +5,8 @@ import { hideBin } from "yargs/helpers";
 
 import { NativeMessagingVersion } from "@bitwarden/common/enums/nativeMessagingVersion";
 
-import { LogUtils } from "../logUtils";
-import NativeMessageService from "../nativeMessageService";
+import { LogUtils } from "../log-utils";
+import NativeMessageService from "../native-message.service";
 import * as config from "../variables";
 
 const argv: any = yargs(hideBin(process.argv)).option("userId", {
