+++ apps/desktop/native-messaging-test-runner/src/log-utils.ts
