+++ apps/desktop/src/main/desktop-credential-storage-listener.ts
