+++ apps/desktop/src/main/native-messaging.main.ts
