+++ apps/desktop/src/main/power-monitor.main.ts
