+++ apps/desktop/src/models/native-messaging/encrypted-command.ts
