+++ apps/desktop/src/models/native-messaging/encrypted-message-payloads/credential-retrieve-payload.ts
