+++ apps/desktop/src/models/native-messaging/encrypted-message-payloads/credential-update-payload.ts
