+++ apps/desktop/src/models/native-messaging/encrypted-message-response.ts
@@ -1,6 +1,6 @@
 import { EncString } from "@bitwarden/common/models/domain/enc-string";
 
-import { MessageCommon } from "./messageCommon";
+import { MessageCommon } from "./message-common";
 
 export type EncryptedMessageResponse = MessageCommon & {
   encryptedPayload: EncString;
