+++ apps/desktop/src/models/native-messaging/encrypted-message-responses/account-status-response.ts
