+++ apps/desktop/src/models/native-messaging/encrypted-message-responses/cannot-decrypt-error-response.ts
