+++ apps/desktop/src/models/native-messaging/encrypted-message-responses/cipher-response.ts
