+++ apps/desktop/src/models/native-messaging/encrypted-message-responses/encrypted-message-response.ts
@@ -1,0 +1,16 @@
+import { AccountStatusResponse } from "./account-status-response";
+import { CannotDecryptErrorResponse } from "./cannot-decrypt-error-response";
+import { CipherResponse } from "./cipher-response";
+import { FailureStatusResponse } from "./failure-status-response";
+import { GenerateResponse } from "./generate-response";
+import { SuccessStatusResponse } from "./success-status-response";
+import { UserStatusErrorResponse } from "./user-status-error-response";
+
+export type EncyptedMessageResponse =
+  | AccountStatusResponse[]
+  | CannotDecryptErrorResponse
+  | CipherResponse[]
+  | FailureStatusResponse
+  | GenerateResponse
+  | SuccessStatusResponse
+  | UserStatusErrorResponse;
