+++ apps/desktop/src/models/native-messaging/encrypted-message-responses/failure-status-response.ts
