+++ apps/desktop/src/models/native-messaging/encrypted-message-responses/generate-response.ts
