+++ apps/desktop/src/models/native-messaging/encrypted-message-responses/success-status-response.ts
