+++ apps/desktop/src/models/native-messaging/encrypted-message-responses/user-status-error-response.ts
