+++ apps/desktop/src/models/native-messaging/index.ts
@@ -1,0 +1,25 @@
+export * from "./encrypted-message-payloads/credential-create-payload";
+export * from "./encrypted-message-payloads/credential-retrieve-payload";
+export * from "./encrypted-message-payloads/credential-update-payload";
+export * from "./encrypted-message-payloads/password-generate-payload";
+
+export * from "./encrypted-message-responses/account-status-response";
+export * from "./encrypted-message-responses/cannot-decrypt-error-response";
+export * from "./encrypted-message-responses/cipher-response";
+export * from "./encrypted-message-responses/encrypted-message-response";
+export * from "./encrypted-message-responses/failure-status-response";
+export * from "./encrypted-message-responses/generate-response";
+export * from "./encrypted-message-responses/success-status-response";
+export * from "./encrypted-message-responses/user-status-error-response";
+
+export * from "./decrypted-command-data";
+export * from "./encrypted-command";
+export * from "./encrypted-message";
+export * from "./encrypted-message-response";
+export * from "./legacy-message";
+export * from "./legacy-message-wrapper";
+export * from "./message";
+export * from "./message-common";
+export * from "./unencrypted-command";
+export * from "./unencrypted-message";
+export * from "./unencrypted-message-response";
