+++ apps/desktop/src/models/native-messaging/legacy-message-wrapper.ts
@@ -1,6 +1,6 @@
 import { EncString } from "@bitwarden/common/models/domain/enc-string";
 
-import { LegacyMessage } from "./legacyMessage";
+import { LegacyMessage } from "./legacy-message";
 
 export type LegacyMessageWrapper = {
   message: LegacyMessage | EncString;
