+++ apps/desktop/src/models/native-messaging/legacy-message.ts
