+++ apps/desktop/src/models/native-messaging/message-common.ts
