+++ apps/desktop/src/models/native-messaging/message.ts
@@ -1,0 +1,4 @@
+import { EncryptedMessage } from "./encrypted-message";
+import { UnencryptedMessage } from "./unencrypted-message";
+
+export type Message = UnencryptedMessage | EncryptedMessage;
