+++ apps/desktop/src/models/native-messaging/unencrypted-command.ts
