--- apps/desktop/src/models/nativeMessaging/index.ts
@@ -1,25 +1,0 @@
-export * from "./encryptedMessagePayloads/credentialCreatePayload";
-export * from "./encryptedMessagePayloads/credentialRetrievePayload";
-export * from "./encryptedMessagePayloads/credentialUpdatePayload";
-export * from "./encryptedMessagePayloads/passwordGeneratePayload";
-
-export * from "./encryptedMessageResponses/accountStatusResponse";
-export * from "./encryptedMessageResponses/cannotDecryptErrorResponse";
-export * from "./encryptedMessageResponses/cipherResponse";
-export * from "./encryptedMessageResponses/encryptedMessageResponse";
-export * from "./encryptedMessageResponses/failureStatusResponse";
-export * from "./encryptedMessageResponses/generateResponse";
-export * from "./encryptedMessageResponses/successStatusResponse";
-export * from "./encryptedMessageResponses/userStatusErrorResponse";
-
-export * from "./decryptedCommandData";
-export * from "./encryptedCommand";
-export * from "./encryptedMessage";
-export * from "./encryptedMessageResponse";
-export * from "./legacyMessage";
-export * from "./legacyMessageWrapper";
-export * from "./message";
-export * from "./messageCommon";
-export * from "./unencryptedCommand";
-export * from "./unencryptedMessage";
-export * from "./unencryptedMessageResponse";
