--- apps/desktop/src/models/nativeMessaging/message.ts
@@ -1,4 +1,0 @@
-import { EncryptedMessage } from "./encryptedMessage";
-import { UnencryptedMessage } from "./unencryptedMessage";
-
-export type Message = UnencryptedMessage | EncryptedMessage;
