+++ apps/desktop/src/services/password-reprompt.service.ts
